Move pg fixture to a separate file
